Move RestAPI module to models.py
