+++ dspyce/rest/__init__.py
@@ -1,3 +1,3 @@
-from dspyce.rest.RestAPI import RestAPI
+from dspyce.rest.models import RestAPI
 from dspyce.rest.functions import authenticate_to_rest
-from dspyce.rest.RestAPI import json_to_object, object_to_json
+from dspyce.rest.models import json_to_object, object_to_json
