+++ dspyce/rest/functions.py
@@ -1,7 +1,7 @@
 from getpass import getpass
 import logging
 import requests
-from dspyce.rest.RestAPI import RestAPI
+from dspyce.rest.models import RestAPI
 
 
 def authenticate_to_rest(rest_api: str, user: str = None, log_level=logging.INFO, log_file: str = None) -> RestAPI:
